Ensure that color bytes are sent in the correct order
It worked fine, before, but this is more robust. Now it doesn't matter what order the
parts of `Color` are defined in (could be {b,r,g} now instead of {b,g,r}), because the
transmitting function guarantees that they'll be sent in the correct order.

If this increases code size or slows things down too much, we can drop this change, of
course.